Separate ERF constructors into separate file; add function which allows us to change from mesh type VariableDz to StretchedDz if terrain is flat
