+++ Source/Make.package
@@ -7,6 +7,7 @@ CEXE_headers += ERF_IndexDefines.H
 CEXE_headers += ERF_Constants.H
 CEXE_sources += ERF_Tagging.cpp
 
+CEXE_sources += ERF_Constructors.cpp
 CEXE_sources += ERF_Diagnostics.cpp
 CEXE_sources += ERF_MakeNewLevel.cpp
 CEXE_sources += ERF_MakeNewArrays.cpp
